work_with_db/requirements.txt
psycopg2
SQLAlchemy
pymongo
python-dotenv